backend/requirements.txt
flask
flask-cors
python-dotenv
pymongo
pydantic
bson
docker
requests
reportlab